requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml>=6.0.1
pydantic==2.5.3
requests==2.31.0

# PDF Processing
requests>=2.31.0

# LLM APIs
openai==1.3.9
anthropic==0.7.8

# Data processing
python-dateutil==2.8.2

# Logging and utilities
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: async support
aiohttp==3.9.1
asyncio==3.4.3

# Code quality
black==23.12.1
flake8==6.1.0
